Mount and uses the REGISTRY_AUTH_FILE when starting a system site
Podman uses the $HOME/.docker/config.json by default, but when using
the Docker socket endpoint, it is not using it, so we need to explicitly
provide the config.json to be used.

This PR is also useful to podman sites, in case the config.json is not
present in the default location, allowing users to pass the
REGISTRY_AUTH_FILE environment variable when executing the system module.

Fixes #72.